Splits the R engine from the R console.
