+++ r-engine/pom.xml
@@ -1,8 +1,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>r-console</artifactId>
-    <name>R console</name>
+    <artifactId>r_engine</artifactId>
+    <name>R engine</name>
     <version>5.1.0</version>
     <url>http://orbisgis.org</url>
     <parent>
@@ -11,7 +11,7 @@
         <version>5.1</version>
         <relativePath>../</relativePath>
     </parent>
-    <description>A R console for OrbisGIS platform based on Renjin.</description>
+    <description>A R engine based on Renjin.</description>
     <packaging>bundle</packaging>
     <build>
         <plugins>
@@ -35,10 +35,25 @@
                 <extensions>true</extensions>
                 <configuration>
                     <instructions>
-                        <Private-Package>org.orbisgis.r.*</Private-Package>
                         <Bundle-Vendor>LabSTICC - CNRS, UMR 6285.</Bundle-Vendor>
-                        <Bundle-Category>OrbisGIS, Console, Language</Bundle-Category>
-                        <Bundle-Icon>org/orbisgis/r/icon.png;size=32</Bundle-Icon>
+                        <Bundle-Icon>org/orbisgis/r_engine/icon.png;size=32</Bundle-Icon>
+                        <Import-Package>
+                            org.renjin.compiler,
+                            org.renjin.datasets,
+                            org.renjin.graphics,
+                            org.renjin.grDevices,
+                            org.renjin.grid,
+                            org.renjin.hamcrest,
+                            org.renjin.methods,
+                            org.renjin.parallel,
+                            org.renjin.splines,
+                            org.renjin.stats,
+                            org.renjin.stats4,
+                            org.renjin.tcltk,
+                            org.renjin.tools,
+                            org.renjin.utils,
+                            *
+                        </Import-Package>
                     </instructions>
                 </configuration>
             </plugin>
@@ -56,8 +71,8 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <poDirectory>${project.build.sourceDirectory}/../resources/org/orbisgis/r/translation/</poDirectory>
-                    <targetBundle>org.orbisgis.r.Messages</targetBundle>
+                    <poDirectory>${project.build.sourceDirectory}/../resources/org/orbisgis/r_engine/translation/</poDirectory>
+                    <targetBundle>org.orbisgis.r_engine.Messages</targetBundle>
                     <keywords>-ktr</keywords>
                     <outputFormat>properties</outputFormat>
                 </configuration>
@@ -78,6 +93,21 @@
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
             <version>1.6.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.orbisgis</groupId>
+            <artifactId>orbisgis-sif</artifactId>
+            <version>5.1.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.renjin</groupId>
+            <artifactId>renjin-script-engine</artifactId>
+            <version>0.9.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.renjin</groupId>
+            <artifactId>renjin-aether-package-loader</artifactId>
+            <version>0.9.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.osgi</groupId>
@@ -101,11 +131,6 @@
             <groupId>${project.groupId}</groupId>
             <artifactId>wkgui-api</artifactId>
             <version>5.1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.orbisgis</groupId>
-            <artifactId>r_engine</artifactId>
-            <version>5.1.0</version>
         </dependency>
     </dependencies>
     <repositories>
